+++ src/main/java/com/tregouet/occam/alg/setters/weighs/properties/impl/BoundVarCounter.java
@@ -26,7 +26,7 @@ public class BoundVarCounter implements PropertyWeigher {
 				}
 			}
 		}
-		property.setWeight((double) boundVars.size());
+		property.setWeight(boundVars.size());
 	}
 
 }
